Implement <value,count> API
When the data set contains a lot of redundancy (i.e. values with many
repetitions), it may be more efficient to add all occurrences to the
t-digest at once instead of adding the values one by one. This reduces
the number of SQL function calls.

We might add one one huge centroid, but that would be against the idea
of having centroids small enough to allow incremental merging. Instead
we generate a t-digest, where all centroids have the same mean but are
sufficiently small to match the rules for well-formed t-digests.

We do that by solving the two inequalities

    z <= q0 * (1 - q0)
    z <= q2 * (1 - q2)

and adding the largest possible centroid in each step. This is can be
done in a single pass, and is much faster than adding the values one by
one, as it does not require any merges/compactions of the t-digest.

This implements an API, originally proposed by Min Wei, with various
improvements and fixes by me.