Implement <value,count> API
When the data set contains a lot of redundancy (i.e. the values repeat),
it may be more efficient to add all occurrences to the t-digest at once
instead of adding the values one by one. This reduces the number of SQL
function calls.

This implements an API, originally proposed by Min Wei, with various
improvements and fixes by me.