Rebalance the sorted centroids
Sorting centroids by (mean, count) works fine, as long as there's enough
variability in the mean value. In some rare cases we can end up with
many centroids having the same mean (e.g. adding the same value many
times, in which case the centroids are sorted mostly by the count in
ascending order. That can have negative impact on percentiles above
median, because it contradicts the idea that we get smaller centroids on
the tails (close to 1.0).

We can solve this by reverting the order of centroids in groups that are
above median, and rebalancing groups that are partially above median.
This is heuristics and experimental, but seems to be working.

This is partially inspired by this commit in Ted Dunning's tdigest
repository:

https://github.com/tdunning/t-digest/commit/eca1125a39c13e918d7054105dc89bcda11cfa44