Represent centroids as (mean,count) instead of (sum,count)
Until now centroids were represented as (sum,count), but that requires
updating the sum even when adding a value matching the mean exactly, or
when merging centroids with the same mean. That leads to rounding
errors, producing unexpected results.

In practice, this does not cause much trouble as the data sets usually
contain many different values, so the chance of matching mean exactly
is fairly rare. But in some cases, e.g. when using the <value,count>
interface, this happens rather often.

This changes the on-disk representation, but in a backwards-compatible
way, i.e. existing values are still interpreted as (sum,count) but new
values are stored as (mean,count).