Fix: mark tdigest_deserialize as strict
The tdigest_serialize function is already marked as STRICT, but the
tdigest_deserialize function was not. That might have led to crashes,
for example with this query:

    with x as (select 123 as z from generate_series(1,100000)
               union all
               select 222 from generate_series(1,100000))
    select tdigest_percentile(x.z, 120, 0.75) from x

This seems to be fairly difficult to trigger - without the UNION ALL,
or with parallel_leader_participation = off the crash goes away.

Fixed by marking tdigest_deserialize as STRICT.