Fix issues in t-digest format update
Commit 8aa6f67aaf switched centroids from (sum,count) to (mean,count)
and implemented automatic conversion to the new format, but there were
about two issues:

- The t-digest was updated in place, but that's wrong - it might be a
  pointer to a buffer, and we should not modify that. We have to copy
  the digest in this case.

- tdigest_out() did not update the format, but then still checked it
  (although only in an assert). We can output the old format just fine
  and we expect tdigest_in() on the other side to understand both the
  old ond new format, so just remove the conversion call.

- tdigest_digest_to_aggstate() did not update the format as it should,
  so fixed by adding the missing tdigest_update_format() call.

It's not clear whether to force tdigest_out/tdigest_send to only output
the new (mean,count) format, or use the actual format of the digest. For
the extension if does not really matter, we can upgrade the format in
output or input function. But maybe producing just the new format would
make it easier to process it in external tools.