Add cast from tdigest to JSON
This commit allows formatting t-digest values as valid JSON documents,
making it easier to use them in JavaScript apps. The default text format
was meant to be human readable, and would require parsing and building
the JSON manually.

We make sure to always print mean, even for tdigests in the older format
storing sum for centroids. Otherwise the "mean" key would be confusing.
But we don't call tdigest_update_format, and instead we simply update
the flags and convert the sum/mean values.

The centroids are stored in two separate arrays - one for means, one for
counts. That makes it easier to process, because it's clear the i-th in
each array is for i-th centroid. We might store it in a single array,
but then we'd have to walk it in pairs. And it'd mix float and int
values in the same array.