Add cast from tdigest to DOUBLE PRECISION[]
This commit allows formatting t-digest values as double precision arrays.

The whole digest is stored in a single "double precision" array, which
may be a bit confusing and perhaps fragile if more fields need to be
added in the future. The initial elements are flags, count (number of
items added to the digest), compression (determines the limit on number
of centroids) and current number of centroids. Follows stream of values
encoding the centroids in pairs of (mean, count).

We make sure to always print mean, even for tdigests in the older format
storing sum for centroids. Otherwise the "mean" key would be confusing.
But we don't call tdigest_update_format, and instead we simply update the
flags and convert the sum/mean values.