Reduce amounts of data in regression tests
The regression tests used fairly large amounts of data (up to 1M of rows
in some cases), so that the whole suite took quite long, particularly
with asserts enabled. On my laptop that took ~7 minutes, which is not
very practical.

This reduces the amounts of data to ~1/10, which reduces the duration of
a test run to about 1/10 too.