Add trimmed aggregates
Implements sum/avg trimmed aggregates, discarding extreme high/low
values determined by the low/high percentiles. The basic aggregate
functions added are

- tdigest_sum(value, low, high)
- tdigest_sum(value, count, low, high)
- tdigest_sum(tdigest, low, high)
- tdigest_avg(value, low, high)
- tdigest_avg(value, count, low, high)
- tdigest_avg(tdigest, low, high)

The patch also adds two regular functions

- tdigest_digest_sum(tdigest, low, high)
- tdigest_digest_avg(tdigest, low, high)

The purpose of those functions is to allow calculating the sum/avg for
individual digests, without having to aggregate the data (which may be
quite inconvenient in some cases).

In all cases the low/high parameters are the quantiles determining the
part of the data to include in the result.