Split regression tests into smaller parts
As the SQL file grew it became a bit cumbersome to work with it, and
due to supporting multiple Postgres versions we may get different
results, requiring alternative expected results. So split the file into
smaller parts, to make it easier to work with, which also makes it
easier to maintain the expected files.

Since Postgres 12 the output seems quite stable, but 9.6, 10 and 11 all
need some changes, due to percentile_disc not being parallel safe,
different handling of CTEs etc.

I've considered reworking the tests so that there are no differences
(e.g. by using plain subqueries instead of CTE), but that does not seem
like a good idea. There'll always be differences that are hard to
eliminate (e.g. due to percentile_disc not being parallel safe), and to
some extent it's actually beneficial to test different plans.